Making configurations for minimum and maximum total scrap value per level usable
The game, as at the moment of this commit, does not use either ``SelectableLevel.minTotalScrapValue`` and ``SelectableLevel.maxTotalScrapValue`` in the logic of spawning scrap in the level, thus rendering any changes of these values in the configurations of each level useless.

In order to make them used, a transpiler patch is required in the spawn scrap method which influences the logic of spawning scrap with a notion of a maximum scrap value it can generate up to in the level.

Fixes #82